refactor: clean coding components folder
