+++ components/AuthCheck/types.ts
@@ -1,0 +1,4 @@
+export interface AuthCheckProps {
+  children: React.ReactNode;
+  fallback?: React.ReactNode;
+}
