+++ components/FormatedTime/types.ts
@@ -1,0 +1,6 @@
+export type TimeInput = number | string;
+
+export type FormatedTimeProps = {
+  time: TimeInput;
+  className?: string;
+};
